Updates to teleop for next competition
We’ve made a bunch of changes to the robot between competitions.
We removed the extend arm motor and replaced it with a servo. Since it is far lighter, it should make climbing possible. This changed some of the button mappings.

We also had a bug where moving forward and backward were reversed.

Also - we moved a standoff at the bottom of the arm to ensure it fits in a 18x18x18 cube. That required the lift arm motor to have new encoder values for the various positions. Since our baskets aren’t necessarily the exact heights, we may need to change these values in the practice arena.