Changes to simple autonomous
We used this simple autonomous to compete in December 2024. It scored a few times, but was inconsistent in responding to sensor values. That is likely because the sampling rate is not sensitive enough. I’ll experiment with fixing that, but for the moment, this code deserves to be checked in.